requirements.txt
streamlit
folium
geopandas
geopy
scipy
pandas
python-dotenv
streamlit-folium
